+++ tests/ttnn/unit_tests/gtests/ccl/test_1d_fabric_loopback_latency.cpp
@@ -466,7 +466,7 @@ inline void RunPersistent1dFabricLatencyTest(
     for (size_t i = 0; i < programs.size(); i++) {
         auto d = devices_with_workers.at(i);
         auto& program = programs.at(i);
-        tt_metal::DumpDeviceProfileResults(d, program);
+        tt_metal::detail::DumpDeviceProfileResults(d, program);
     }
     log_info(tt::LogTest, "Finished");
 }
