+++ tt_metal/programming_examples/profiler/test_dispatch_cores/test_dispatch_cores.cpp
@@ -52,7 +52,7 @@ void RunCustomCycle(tt_metal::IDevice* device, int loop_count) {
         tt_metal::ComputeConfig{.compile_args = trisc_kernel_args, .defines = kernel_defines});
 
     EnqueueProgram(device->command_queue(), program, false);
-    tt_metal::DumpDeviceProfileResults(device, program);
+    tt_metal::detail::DumpDeviceProfileResults(device, program);
 }
 
 int main() {
